requirements.txt
Pillow==11.0.0
python-barcode==0.14.0
pywin32>=307
keyboard==0.13.5
requests==2.31.0
pyserial==3.5 